requirements-container.txt
grpcio
grpcio-tools
PyAstronomy~=0.16.0
numpy~=1.20.3
vtk~=9.0.1
networkx~=2.5.1
scipy==1.6.3
matplotlib==3.4.2
flask==2.0.0